SEP6-99: add MovieListMovieService

+++ src/MovieManagement.Functions/Services/MovieListService.cs
@@ -3,14 +3,12 @@
 public class MovieListService : IMovieListService
 {
     private readonly IMovieListRepository _repository;
-    private readonly IMovieListMovieRepository _listMovieRepository;
     private readonly IUserRepository _userRepository;
     private readonly IMapper _mapper;
 
-    public MovieListService(IMovieListRepository repository, IMapper mapper, IMovieListMovieRepository listMovieRepository, IUserRepository userRepository) {
+    public MovieListService(IMovieListRepository repository, IMapper mapper, IUserRepository userRepository) {
         _repository = repository;
         _mapper = mapper;
-        _listMovieRepository = listMovieRepository;
         _userRepository = userRepository;
     }
 
@@ -18,9 +16,6 @@ public class MovieListService : IMovieListService
 
         var movieListEntity = _mapper.Map<MovieListEntity>(addMovieListDto);
         movieListEntity.MovieListId = new Guid();
-        if ((await _userRepository.GetAsync(addMovieListDto.UserId)) is null) {
-            throw new Exception("User doesn't exist");
-        }
 
         await _repository.AddAsync(movieListEntity);
 
@@ -28,31 +23,7 @@ public class MovieListService : IMovieListService
         return movieList;
     }
 
-    public async Task<MovieListDto> AddMovieToMovieListAsync(MovieToMovieListDto movieToMovieListDto)
-    {
-        var movieListMovieEntity = _mapper.Map<MovieListMovieEntity>(movieToMovieListDto);
-        if (await _repository.GetAsync(movieListMovieEntity.MovieListId) is null)
-        {
-            throw new Exception("Movie list doesn't exist");
-        }
-        
-        var existingMovieInList = await _listMovieRepository.GetMovieFromMovieListAsync(movieListMovieEntity);
-        
-        if (existingMovieInList != null)
-        {
-            throw new Exception("The movie already exists in the list");
-        }
-        
-        await _listMovieRepository.AddMovieToMovieListAsync(movieListMovieEntity);
-       
-        var listEntity = await _repository.GetAsync(movieToMovieListDto.MovieListId);
-        var movies = await _listMovieRepository.GetMoviesByListAsync(movieToMovieListDto.MovieListId);
-        
-        var movieList = _mapper.Map<MovieListDto>(listEntity);
-        movieList.Movies = _mapper.Map<List<MovieDto>>(movies);
-        
-        return movieList;
-    }
+   
 
     public async Task<List<MovieListDto>> GetMovieListsAsync(Guid userId) {
         if ((await _userRepository.GetAsync(userId)) is null) {
@@ -61,7 +32,7 @@ public class MovieListService : IMovieListService
         var list = await _repository.GetMovieListsByUserAsync(userId);
         var mappedList = _mapper.Map<List<MovieListDto>>(list);
         foreach (var movieList in mappedList) {
-            var movies = await _listMovieRepository.GetMoviesByListAsync(movieList.MovieListId);
+            var movies = await _repository.GetMoviesByListAsync(movieList.MovieListId);
             movieList.Movies = _mapper.Map<List<MovieDto>>(movies);
         }
         return mappedList;
@@ -73,15 +44,12 @@ public class MovieListService : IMovieListService
             throw new Exception("Movie List doesn't exist");
         }
 
+        if (movieList.Title.Equals("ToWatch") || movieList.Title.Equals("Favourites")) {
+            throw new Exception("Cannot delete movie list");
+        }
+
         await _repository.DeleteAsync(movieListId);
     }
 
-    public async Task DeleteMovieFromMovieListAsync(MovieToMovieListDto movieToMovieListDto) {
-        var entity = _mapper.Map<MovieListMovieEntity>(movieToMovieListDto);
-        if (await _listMovieRepository.GetMovieFromMovieListAsync(entity) is null) {
-            throw new Exception("Movie not in list");
-        }
-
-        await _listMovieRepository.DeleteMovieFromMovieListAsync(entity);
-    }
+    
 }
